feat(clear-empty-dir): 添加试运行功能

+++ scripts/filesystem/clear_empty_dir.nu
@@ -2,20 +2,25 @@ use internal.nu *
 
 # 删除一个目录下的空目录
 export def clear-empty-dir [
-  dir: path
-  --recursion(-r) # 是否递归删除
-  --verbose(-v) # 是否显示删除的目录
+  dir?: path
+  --recursion(-r) # 递归删除
+  --verbose(-v)   # 显示删除的目录
+  --dry-run(-n)   # 试运行
 ] : [
   nothing -> nothing
 ] {
+  let dir = if ($dir | is-empty) { pwd } else { $dir }
   if $recursion {
-    clear-empty-dir-recursion --verbose=$verbose $dir
+    clear-empty-dir-recursion --verbose=$verbose --dry-run=$dry_run $dir
   } else {
     get-empty-dir $dir | each {|it|
       if $verbose {
         print $"deleted ($it)"
       }
-      rm $it
+      if not $dry_run {
+        rm $it
+      }
     }
+    | ignore
   }
 }
